Update README.md
The bus app ticket booking system was developed as a college project by students of Softwarica College, a leading IT college in Nepal. The project was aimed at providing a solution to the challenges faced by travelers when booking bus tickets, such as long queues and the need to physically visit ticket offices.

The students utilized their knowledge and skills in software development to create a user-friendly application that simplifies the bus ticket booking process. They incorporated various features such as real-time updates and secure payment options, ensuring a seamless booking experience for users.